requirements.txt
anyio==3.6.2
attrs==22.1.0
certifi==2022.9.24
charset-normalizer==2.1.1
exceptiongroup==1.0.0rc9
fastapi==0.85.1
greenlet==1.1.3.post0
idna==3.4
iniconfig==1.1.1
Mako==1.2.3
MarkupSafe==2.1.1
packaging==21.3
pluggy==1.0.0
psycopg2-binary==2.9.5
pydantic==1.10.2
pyparsing==3.0.9
pytest==7.2.0
requests==2.28.1
sniffio==1.3.0
SQLAlchemy==1.4.42
starlette==0.20.4
tomli==2.0.1
typing_extensions==4.4.0
urllib3==1.26.12
